requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
lxml>=4.9.0
selenium>=4.15.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
